\name{TCGAMethylation450k-package}
\alias{TCGAMethylation450k-package}
\alias{TCGAMethylation450k}
\docType{package}
\title{
  The Cancer Genome Atlas Illumina 450k methylation example data
}
\description{
The Cancer Genome Atlas (TCGA) is applying genomics technologies
to over 20 different types of cancer.  This package contains a small set of
450k array data in idat format.

Data are included in this package in extdata/idat.  A typical
way of accessing these data would be to use the methylumi package
methylumIDAT function.
}
\details{
\tabular{ll}{
Package: \tab TCGAMethylation450k\cr
Type: \tab Package\cr
License: \tab GPL-2\cr
}

}
\author{
Sean Davis

Maintainer: Sean Davis <sdavis2@mail.nih.gov>
}
\references{
  \url{http://cancergenome.nih.gov/}
}
\keyword{ package }
\seealso{
\code{\link[methylumi:methylumi-package]{methylumi}}
}
